define: memo_utils.jsを定義、memo.js内のメソッドを移動
